Education API with basic CRUD finished

+++ Web/MySkillsServer.Web/Controllers/EducationsController.cs
@@ -1,5 +1,6 @@
 ﻿namespace MySkillsServer.Web.Controllers
 {
+    using System.Collections.Generic;
     using System.Linq;
     using System.Threading.Tasks;
 
@@ -25,59 +26,82 @@
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IEnumerable<EducationExportModel>> GetAll()
         {
-            var model = await this.educationsService.GetAllAsNoTrackingOrderedAsync<EducationExportModel>();
+            var models = await this.educationsService.GetAllAsNoTrackingOrderedAsync<EducationExportModel>();
 
-            return (IActionResult)model;
+            return models;
         }
 
         // GET /api/educations/id and api/educations?id=1234
         [HttpGet("{id}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(404)]
-        public async Task<IActionResult> GetById(int id)
+        public async Task<ActionResult<EducationExportModel>> GetById(int id)
         {
             var model = await this.educationsService.GetByIdAsync<EducationExportModel>(id);
+
             if (model == null)
             {
                 return this.NotFound();
             }
 
-            return (IActionResult)model;
+            return model;
         }
 
         [HttpPost]
+        [IgnoreAntiforgeryTokenAttribute]
         // [Authorize]
-        public async Task<IActionResult> Create(EducationEditInputModel input)
+        public async Task<ActionResult<EducationCreateInputModel>> Create(EducationCreateInputModel input)
         {
-            // var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
-            var user = await this.userManager.GetUserAsync(this.User);
+            //var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            //var user = await this.userManager.GetUserAsync(this.User);
 
-            await this.educationsService.CreateAsync(input, user.Id);
+            //await this.educationsService.CreateAsync(input, user.Id);
+            await this.educationsService.CreateAsync(input);
 
-            return (IActionResult)input;
+            return input;
         }
 
-        [HttpPatch]
+        [HttpPut("{id}")]
+        [IgnoreAntiforgeryTokenAttribute]
         // [Authorize]
-        public async Task<IActionResult> Edit(EducationEditInputModel input)
+        public async Task<ActionResult<EducationEditInputModel>> Edit(int id, EducationEditInputModel input)
         {
+            if (id != input.Id)
+            {
+                return this.BadRequest();
+            }
+
+            var model = await this.educationsService.GetByIdAsync<EducationExportModel>(id);
+
+            if (model == null)
+            {
+                return this.NotFound();
+            }
+
             // var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
-            var user = await this.userManager.GetUserAsync(this.User);
+            //var user = await this.userManager.GetUserAsync(this.User);
 
-            await this.educationsService.CreateAsync(input, user.Id);
+            //await this.educationsService.EditAsync(input, user.Id);
+            await this.educationsService.EditAsync(input);
 
-            return (IActionResult)input;
+            return this.NoContent();
         }
 
         [HttpDelete("{id}")]
+        [IgnoreAntiforgeryTokenAttribute]
         // [Authorize]
-        public int Delete(int id)
+        public async Task<ActionResult<int>> Delete(int id)
         {
-            this.educationsService.DeleteAsync(id);
+            var model = await this.educationsService.GetByIdAsync<EducationExportModel>(id);
 
-            return id;
+            if (model == null)
+            {
+                return this.NotFound();
+            }
+
+            return await this.educationsService.DeleteAsync(id);
         }
     }
 }
